small change of code due to interface change after rebase

+++ integrals/atomic_integral.h
@@ -13,6 +13,7 @@
 #include "../utility/wcout_utf8.h"
 #include "../ta_routines/array_to_eigen.h"
 #include "../utility/parallel_print.h"
+#include "../utility/time.h"
 
 namespace mpqc{
 namespace integrals{
@@ -143,23 +144,23 @@ namespace integrals{
                 if(operation.oper() == Operation::Operations::cGTG2){
 
                     auto squared_pragmas = f12::gtg_params_squared(this->gtg_params_);
-                    libint2::TwoBodyEngine<libint2::cGTG> engine(max_nprim, static_cast<int>(max_am),0,std::numeric_limits<double>::epsilon(),squared_pragmas);
+                    libint2::TwoBodyEngine<libint2::cGTG> engine(max_nprim, max_am,0,std::numeric_limits<double>::epsilon(),squared_pragmas);
                     auto engine_pool = make_pool(engine);
                     result = compute_integrals(this->world_,engine_pool,bs_array);
 
                 }else{
-                    libint2::TwoBodyEngine<libint2::cGTG> engine(max_nprim, static_cast<int>(max_am),0,std::numeric_limits<double>::epsilon(),this->gtg_params_);
+                    libint2::TwoBodyEngine<libint2::cGTG> engine(max_nprim, max_am,0,std::numeric_limits<double>::epsilon(),this->gtg_params_);
                     auto engine_pool = make_pool(engine);
                     result = compute_integrals(this->world_,engine_pool,bs_array);
                 }
             }
             else if(kernel == libint2::cGTG_times_Coulomb){
-                libint2::TwoBodyEngine<libint2::cGTG_times_Coulomb> engine(max_nprim, static_cast<int>(max_am),0,std::numeric_limits<double>::epsilon(),this->gtg_params_);
+                libint2::TwoBodyEngine<libint2::cGTG_times_Coulomb> engine(max_nprim,max_am,0,std::numeric_limits<double>::epsilon(),this->gtg_params_);
                 auto engine_pool = make_pool(engine);
                 result = compute_integrals(this->world_,engine_pool,bs_array);
             }
             else if(kernel == libint2::DelcGTG_square){
-                libint2::TwoBodyEngine<libint2::DelcGTG_square> engine(max_nprim, static_cast<int>(max_am),0,std::numeric_limits<double>::epsilon(),this->gtg_params_);
+                libint2::TwoBodyEngine<libint2::DelcGTG_square> engine(max_nprim, max_am,0,std::numeric_limits<double>::epsilon(),this->gtg_params_);
                 auto engine_pool = make_pool(engine);
                 result = compute_integrals(this->world_,engine_pool,bs_array);
             }
